stable-diffusion/keras-example/requirements.txt
keras-cv
tensorflow_datasets